requirements.txt
uvicorn==0.34.0,<0.35
fastapi==0.115.12,<0.116
itsdangerous==2.2.0,<2.3
python-dotenv==1.1.0,<1.2
loguru==0.7.3,<0.8
pinecone-client==6.0.0,<6.1
pydantic==2.11.2,<2.12
pydantic_core==2.33.1,<2.34
pinecone-plugin-interface==0.0.7,   <0.0.8
SQLAlchemy==2.0.40,<2.1
jwt==1.3.1,<1.4
bcrypt==4.3.0,<4.4
pydantic-settings==2.8.1,<2.9
psycopg2==2.9.10,<2.10
python-multipart==0.0.20,<0.1
PyJWT==2.10.1,<2.11